Inferno works! Though with a caveat.
Since we don't have a version of the Inferno sprite that faces up, I had to rotate it using predraw. This however, does not rotate the hitbox, and neither does Projectile.rotation. The result is that the hitbox is much wider than expected, and the solution would be to use a proper projectile sprite.