wip chargeable arrow rework
I want to hold the actual arrow so we can call onkill and onhitnpc in the subclass. But this shit is so fucked. The arrow's position was off so I had to correct it with a few vectors, and now the arrow is super wide and always has the same hitbox shape while flying. It also damages things above it while held for god knows what reason.